+++ driver/src/main/scala/com/stratio/sparta/driver/SpartaClusterJob.scala
@@ -20,15 +20,15 @@ import akka.actor.{ActorSystem, Props}
 import com.google.common.io.BaseEncoding
 import com.stratio.sparta.driver.exception.DriverException
 import com.stratio.sparta.driver.service.StreamingContextService
+import com.stratio.sparta.serving.core.actor.StatusActor
 import com.stratio.sparta.serving.core.actor.StatusActor.Update
-import com.stratio.sparta.serving.core.actor.{FragmentActor, StatusActor}
 import com.stratio.sparta.serving.core.config.SpartaConfig
 import com.stratio.sparta.serving.core.constants.AkkaConstant
 import com.stratio.sparta.serving.core.curator.CuratorFactoryHolder
-import com.stratio.sparta.serving.core.helpers.{FragmentsHelper, ResourceManagerLinkHelper}
+import com.stratio.sparta.serving.core.helpers.ResourceManagerLinkHelper
 import com.stratio.sparta.serving.core.models.enumerators.PolicyStatusEnum._
 import com.stratio.sparta.serving.core.models.policy.{PhaseEnum, PolicyErrorModel, PolicyStatusModel}
-import com.stratio.sparta.serving.core.utils.{PluginsFilesUtils, PolicyConfigUtils, PolicyUtils}
+import com.stratio.sparta.serving.core.utils.{FragmentUtils, PluginsFilesUtils, PolicyConfigUtils, PolicyUtils}
 import com.typesafe.config.ConfigFactory
 import org.apache.curator.framework.CuratorFramework
 
@@ -69,8 +69,10 @@ object SpartaClusterJob extends PluginsFilesUtils {
         val policyUtils = new PolicyUtils {
           override val curatorFramework: CuratorFramework = curatorInstance
         }
-        val fragmentActor = system.actorOf(Props(new FragmentActor(curatorInstance)), AkkaConstant.FragmentActor)
-        val policy = FragmentsHelper.getPolicyWithFragments(policyUtils.getPolicyById(policyId), fragmentActor)
+        val fragmentUtils = new FragmentUtils {
+          override val curatorFramework: CuratorFramework = curatorInstance
+        }
+        val policy = fragmentUtils.getPolicyWithFragments(policyUtils.getPolicyById(policyId))
         val startingInfo = s"Starting policy in cluster"
         log.info(startingInfo)
         statusActor ! Update(PolicyStatusModel(id = policyId, status = Starting, statusInfo = Some(startingInfo)))