+++ serving-api/src/main/scala/com/stratio/sparta/serving/api/actor/PluginActor.scala
@@ -13,16 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.stratio.sparta.serving.api.actor
 
-import java.io.{BufferedOutputStream, FileOutputStream}
-import java.util.function.Predicate
-import java.util.regex.Pattern
-
 import akka.actor.Actor
-import akka.event.slf4j.SLF4JLogging
-import com.stratio.sparta.serving.api.actor.PluginActor.{PluginResponse, UploadFile}
-import com.stratio.sparta.serving.api.constants.HttpConstant
+import com.stratio.sparta.serving.api.actor.PluginActor._
+import com.stratio.sparta.serving.api.utils.FileActorUtils
 import com.stratio.sparta.serving.core.config.SpartaConfig
 import com.stratio.sparta.serving.core.constants.AppConstant
 import com.stratio.sparta.serving.core.models.SpartaSerializer
@@ -31,55 +27,41 @@ import spray.httpx.Json4sJacksonSupport
 
 import scala.util.{Failure, Try}
 
-class PluginActor extends Actor
-  with Json4sJacksonSupport
-  with SLF4JLogging
-  with SpartaSerializer {
+class PluginActor extends Actor with Json4sJacksonSupport with FileActorUtils with SpartaSerializer {
 
   //The dir where the jars will be saved
   val targetDir = Try(SpartaConfig.getDetailConfig.get.getString(AppConstant.PluginsPackageLocation))
     .getOrElse(AppConstant.DefaultPluginsPackageLocation)
-  //Configuration of the app
-  val host = SpartaConfig.apiConfig.get.getString("host")
-  val port = SpartaConfig.apiConfig.get.getInt("port")
-  val targetUrl = s"$host:$port"
-  //Url of the download endpoint
-  val url = s"$targetUrl/${HttpConstant.PluginsPath}"
-  //Regexp for jar name validation
-  val jarFileName: Predicate[String] = Pattern.compile(""".*\.jar""").asPredicate()
 
   override def receive: Receive = {
-    case UploadFile(_, files) if files.isEmpty =>
-      sender ! PluginResponse(Failure(new IllegalArgumentException(s"A file is expected")))
-    case UploadFile(_, files) if files.size != 1 =>
-      sender ! PluginResponse(Failure(new IllegalArgumentException(s"More than one file is not supported")))
-    case UploadFile(name, files) =>
-      if (jarFileName.test(name)) uploadFile(name, files)
-      else sender ! PluginResponse(Failure(new IllegalArgumentException(s"$name is Not a valid file name")))
+    case UploadPlugins(files) =>
+      if (files.isEmpty) sender ! PluginResponse(Failure(new IllegalArgumentException(s"Almost one file is expected")))
+      else uploadPlugins(files)
+    case ListPlugins => browsePlugins()
+    case DeletePlugins => deletePlugins()
+    case DeletePlugin(fileName) => deletePlugin(fileName)
     case _ => log.info("Unrecognized message in Plugin Actor")
   }
 
-  def uploadFile(fileName: String, files: Seq[BodyPart]): Unit = {
-    sender ! PluginResponse(
-      Try {
-        saveFile(files.head.entity.data.toByteArray, s"$targetDir/$fileName")
-        s"$url/$fileName"
-      }
-    )
-  }
+  def deletePlugins(): Unit = sender ! PluginResponse(deleteFiles())
 
-  protected def saveFile(array: Array[Byte], fileName: String): Unit = {
-    log.info(s"Saving file to: $fileName")
-    val bos = new BufferedOutputStream(new FileOutputStream(fileName))
-    bos.write(array)
-    bos.close()
-  }
+  def deletePlugin(fileName: String): Unit = sender ! PluginResponse(deleteFile(fileName))
+
+  def browsePlugins(): Unit = sender ! PluginResponse(browseDirectory())
+
+  def uploadPlugins(files: Seq[BodyPart]): Unit = sender ! PluginResponse(uploadFiles(files))
 }
 
 object PluginActor {
 
-  case class UploadFile(t: String, files: Seq[BodyPart])
+  case class UploadPlugins(files: Seq[BodyPart])
 
   case class PluginResponse(status: Try[_])
+
+  case object ListPlugins
+
+  case object DeletePlugins
+
+  case class DeletePlugin(fileName: String)
 
 }