+++ serving-core/src/main/scala/com/stratio/sparta/serving/core/config/SpartaConfig.scala
@@ -17,133 +17,86 @@
 package com.stratio.sparta.serving.core.config
 
 import akka.event.slf4j.SLF4JLogging
-import com.stratio.sparta.serving.core.constants.AppConstant
+import com.stratio.sparta.serving.core.constants.AppConstant._
 import com.typesafe.config.{Config, ConfigFactory}
 
-import scala.util.{Failure, Success, Try}
+import scala.util.{Success, Try}
 
 /**
- * Helper with common operations used to create a Sparta context used to run the application.
- */
+  * Helper with common operations used to create a Sparta context used to run the application.
+  */
 object SpartaConfig extends SLF4JLogging {
 
   var mainConfig: Option[Config] = None
   var apiConfig: Option[Config] = None
-  var sprayConfig: Option[Config] = None
 
-  /**
-   * Initializes base configuration.
-   *
-   * @param currentConfig if it is setted the function tries to load a node from a loaded config.
-   * @param node          with the node needed to load the configuration.
-   * @return the loaded configuration.
-   */
   def initConfig(node: String,
                  currentConfig: Option[Config] = None,
-                 configFactory: ConfigFactory = new SpartaConfigFactory): Option[Config] = {
+                 configFactory: SpartaConfigFactory = SpartaConfigFactory()): Option[Config] = {
     log.info(s"Loading $node configuration")
     val configResult = currentConfig match {
-      case Some(config) => Try(config.getConfig(node)) match {
-        case Success(config) => Some(config)
-        case _ => None
-      }
-      case _ => configFactory.getConfig(node)
+      case Some(config) => Try(config.getConfig(node)).toOption
+      case None => configFactory.getConfig(node)
     }
     assert(configResult.isDefined, s"Fatal Error: configuration can not be loaded: $node")
     configResult
   }
 
   def initMainConfig(currentConfig: Option[Config] = None,
-                     configFactory: ConfigFactory = new SpartaConfigFactory): Option[Config] = {
-    mainConfig = initConfig(AppConstant.ConfigAppName, currentConfig, configFactory)
+                     configFactory: SpartaConfigFactory = SpartaConfigFactory()): Option[Config] = {
+    mainConfig = initConfig(ConfigAppName, currentConfig, configFactory)
     mainConfig
   }
 
-  def initApiConfig(configFactory: ConfigFactory = new SpartaConfigFactory): Option[Config] = {
-    apiConfig = initConfig(AppConstant.ConfigApi, mainConfig, configFactory)
+  def initApiConfig(configFactory: SpartaConfigFactory = SpartaConfigFactory()): Option[Config] = {
+    apiConfig = initConfig(ConfigApi, mainConfig, configFactory)
     apiConfig
-  }
-
-  def initSprayConfig(configFactory: ConfigFactory = new SpartaConfigFactory): Option[Config] = {
-    sprayConfig = configFactory.getConfig(AppConstant.ConfigSpray)
-    sprayConfig
   }
 
   def getClusterConfig(executionMode: Option[String] = None): Option[Config] =
     Try {
       executionMode match {
         case Some(execMode) if execMode.nonEmpty => execMode
-        case _ => getDetailConfig.get.getString(AppConstant.ExecutionMode)
+        case _ => getDetailConfig.get.getString(ExecutionMode)
       }
     } match {
-      case Success(execMode) =>
-        if (execMode != AppConstant.ConfigLocal) getOptionConfig(execMode, mainConfig.get)
-        else None
-      case Failure(exception) =>
-        log.error("Error when extracting cluster configuration. ", exception)
+      case Success(execMode) if execMode != ConfigLocal =>
+        getOptionConfig(execMode, mainConfig.get)
+      case _ =>
+        log.error("Error when extracting cluster configuration")
         None
     }
 
-  def getHdfsConfig: Option[Config] = mainConfig match {
-    case Some(config) => getOptionConfig(AppConstant.ConfigHdfs, config)
-    case None => None
-  }
+  def getHdfsConfig: Option[Config] = mainConfig.flatMap(config => getOptionConfig(ConfigHdfs, config))
 
-  def getDetailConfig: Option[Config] = mainConfig match {
-    case Some(config) => getOptionConfig(AppConstant.ConfigDetail, config)
-    case None => None
-  }
+  def getDetailConfig: Option[Config] = mainConfig.flatMap(config => getOptionConfig(ConfigDetail, config))
 
-  def getZookeeperConfig: Option[Config] = mainConfig match {
-    case Some(config) => getOptionConfig(AppConstant.ConfigZookeeper, config)
-    case None => None
-  }
+  def getZookeeperConfig: Option[Config] = mainConfig.flatMap(config => getOptionConfig(ConfigZookeeper, config))
 
-  /**
-   * Initializes base configuration.
-   *
-   * @param currentConfig if it is setted the function tries to load a node from a loaded config.
-   * @param node          with the node needed to load the configuration.
-   * @return the optional loaded configuration.
-   */
+  def getSprayConfig: Option[Config] = SpartaConfigFactory().getConfig(ConfigSpray)
+
   def initOptionalConfig(node: String,
                          currentConfig: Option[Config] = None,
-                         configFactory: ConfigFactory = new SpartaConfigFactory): Option[Config] = {
-    log.info(s"> Loading $node configuration")
+                         configFactory: SpartaConfigFactory = SpartaConfigFactory()): Option[Config] = {
+    log.info(s" Loading $node configuration")
     Try(
       currentConfig match {
         case Some(config) => getOptionConfig(node, config)
-        case _ => configFactory.getConfig(node)
+        case None => configFactory.getConfig(node)
       }
     ).getOrElse(None)
   }
 
   def getOptionConfig(node: String, currentConfig: Config): Option[Config] =
-    Try(currentConfig.getConfig(node)) match {
-      case Success(config) => Some(config)
-      case _ => None
-    }
+    Try(currentConfig.getConfig(node)).toOption
 
   def getOptionStringConfig(node: String, currentConfig: Config): Option[String] =
-    Try(currentConfig.getString(node)) match {
-      case Success(config) => Some(config)
-      case _ => None
-    }
-
-  def isHttpsEnabled(): Boolean =
-    SpartaConfig.sprayConfig match {
-      case Some(config) =>
-        Try(config.getValue("ssl-encryption")) match {
-          case Success(value) =>
-            if ("on".equals(value.unwrapped())) true else false
-          case Failure(exception) => false
-        }
-      case None => false
-    }
+    Try(currentConfig.getString(node)).toOption
 
   def daemonicAkkaConfig: Config = mainConfig match {
     case Some(mainSpartaConfig) =>
       mainSpartaConfig.withFallback(ConfigFactory.load(ConfigFactory.parseString("akka.daemonic=on")))
     case None => ConfigFactory.load(ConfigFactory.parseString("akka.daemonic=on"))
   }
+
 }